Add workaround for generating code containing operator<=>.
`operator<=>` is a C++ feature introduced in C++20, which isn't
currently supported by CppSharp. Adding a try-catch block that
skips code generation for the spaceship operator is a minimally
intrusive change that can be removed once spaceship operator code
generation is added to CppSharp.

This workaround is implemented using a try-catch instead of checking
for the operator at the start of the function because this workaround
will be automatically disabled if CppSharp adds support for the
spaceship operator because the exception the block is for will no
longer be thrown.